Landing files refactoring
