+++ src/app/components/landing/Hero.tsx
