+++ src/jest.config.js
@@ -26,8 +26,8 @@ const customJestConfig = {
         '**/*{js,jsx,ts,tsx}',
         '!**/app/api/**',
         '!**/app/layout.tsx',
-        '!**/app/components/Navbars/MobileHeader.tsx',
-        '!**/app/components/Navbars/SideNav.tsx',
+        '!**/app/components/navbars/MobileHeader.tsx',
+        '!**/app/components/navbars/SideNav.tsx',
         '!**/app/layout.tsx',
         '!**/app/constants/types.ts',
         '!**/app/dashboard/layout.tsx',
